+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/projects/ProjectsController.java
@@ -124,10 +124,7 @@ public class ProjectsController {
 
 	// CONSTANTS
 	private final List<Map<String, String>> EXPORT_TYPES = ImmutableList.of(
-			ImmutableMap.of("format", "xlsx", "name", "Excel"), ImmutableMap.of("format", "csv", "name", "CSV")
-	);
-
-
+			ImmutableMap.of("format", "xlsx", "name", "Excel"), ImmutableMap.of("format", "csv", "name", "CSV"));
 
 	@Autowired
 	public ProjectsController(ProjectService projectService, SampleService sampleService, UserService userService,
@@ -152,9 +149,7 @@ public class ProjectsController {
 	/**
 	 * Request for the page to display a list of all projects available to the currently logged in user.
 	 *
-	 * @param model
-	 * 		The model to add attributes to for the template.
-	 *
+	 * @param model The model to add attributes to for the template.
 	 * @return The name of the page.
 	 */
 	@RequestMapping("/projects")
@@ -169,9 +164,7 @@ public class ProjectsController {
 	/**
 	 * Get the admin projects page.
 	 *
-	 * @param model
-	 * 		{@link Model}
-	 *
+	 * @param model {@link Model}
 	 * @return The name of the page
 	 */
 	@RequestMapping("/projects/all")
@@ -186,13 +179,9 @@ public class ProjectsController {
 	/**
 	 * Request for a specific project details page.
 	 *
-	 * @param projectId
-	 * 		The id for the project to show details for.
-	 * @param model
-	 * 		Spring model to populate the html page.
-	 * @param principal
-	 * 		a reference to the logged in user.
-	 *
+	 * @param projectId The id for the project to show details for.
+	 * @param model     Spring model to populate the html page.
+	 * @param principal a reference to the logged in user.
 	 * @return The name of the project details page.
 	 */
 	@RequestMapping(value = "/projects/{projectId}/activity")
@@ -208,18 +197,18 @@ public class ProjectsController {
 	/**
 	 * Gets the name of the template for the new project page
 	 *
-	 * @param useCartSamples
-	 *            Whether or not to use the samples in the cart when creating
-	 *            the project
-	 * @param model
-	 *            {@link Model}
-	 *
+	 * @param useCartSamples Whether or not to use the samples in the cart when creating
+	 *                       the project
+	 * @param model          {@link Model}
+	 * @param owner          whether or not to lock the sample(s) from being modified from new
+	 *                       the project
 	 * @return The name of the create new project page
 	 */
 	@RequestMapping(value = "/projects/new", method = RequestMethod.GET)
 	public String getCreateProjectPage(
 			@RequestParam(name = "cart", required = false, defaultValue = "false") boolean useCartSamples,
-			final Model model) {
+			final Model model,
+			@RequestParam(name = "lockSamples", required = false, defaultValue = "true") boolean owner) {
 		model.addAttribute("useCartSamples", useCartSamples);
 
 		Map<Project, List<Sample>> selected = cartController.getSelected();
@@ -228,15 +217,17 @@ public class ProjectsController {
 		Set<Sample> allowed = new HashSet<>();
 		Set<Sample> disallowed = new HashSet<>();
 
-		selected.values().forEach(set -> {
-			set.stream().forEach(s -> {
-				if (canModifySample(s)) {
-					allowed.add(s);
-				} else {
-					disallowed.add(s);
-				}
-			});
-		});
+		selected.values()
+				.forEach(set -> {
+					set.stream()
+							.forEach(s -> {
+								if (canModifySample(s)) {
+									allowed.add(s);
+								} else {
+									disallowed.add(s);
+								}
+							});
+				});
 
 		model.addAttribute("allowedSamples", allowed);
 		model.addAttribute("disallowedSamples", disallowed);
@@ -250,15 +241,14 @@ public class ProjectsController {
 	/**
 	 * Get the page to synchronize remote projects
 	 *
-	 * @param model
-	 *            Model to render for view
+	 * @param model Model to render for view
 	 * @return Name of the project sync page
 	 */
 	@RequestMapping(value = "/projects/synchronize", method = RequestMethod.GET)
 	public String getSynchronizeProjectPage(final Model model) {
 
 		Iterable<RemoteAPI> apis = remoteApiService.findAll();
-		model.addAttribute("apis",apis);
+		model.addAttribute("apis", apis);
 		model.addAttribute("frequencies", ProjectSyncFrequency.values());
 		model.addAttribute("defaultFrequency", ProjectSyncFrequency.WEEKLY);
 
@@ -285,7 +275,8 @@ public class ProjectsController {
 		try {
 			Project read = projectRemoteService.read(url);
 			read.setId(null);
-			read.getRemoteStatus().setSyncStatus(SyncStatus.MARKED);
+			read.getRemoteStatus()
+					.setSyncStatus(SyncStatus.MARKED);
 			read.setSyncFrequency(syncFrequency);
 
 			read = projectService.create(read);
@@ -308,8 +299,7 @@ public class ProjectsController {
 	 * List all the {@link Project}s that can be read for a user from a given
 	 * {@link RemoteAPI}
 	 *
-	 * @param apiId
-	 *            the local ID of the {@link RemoteAPI}
+	 * @param apiId the local ID of the {@link RemoteAPI}
 	 * @return a List of {@link Project}s
 	 */
 	@RequestMapping(value = "/projects/ajax/api/{apiId}")
@@ -318,42 +308,48 @@ public class ProjectsController {
 		RemoteAPI api = remoteApiService.read(apiId);
 		List<Project> listProjectsForAPI = projectRemoteService.listProjectsForAPI(api);
 
-		return listProjectsForAPI.stream().map(ProjectByApiResponse::new).collect(Collectors.toList());
+		return listProjectsForAPI.stream()
+				.map(ProjectByApiResponse::new)
+				.collect(Collectors.toList());
 	}
 
 	/**
 	 * Creates a new project and displays a list of users for the user to add to
 	 * the project
 	 *
-	 * @param model
-	 *            {@link Model}
-	 * @param project
-	 *            the {@link Project} to create
-	 * @param useCartSamples
-	 *            add all samples in the cart to the project
-	 *
+	 * @param model          {@link Model}
+	 * @param project        the {@link Project} to create
+	 * @param useCartSamples add all samples in the cart to the project
+	 * @param owner          lock sample modification from the new project
 	 * @return The name of the add users to project page
 	 */
 	@RequestMapping(value = "/projects/new", method = RequestMethod.POST)
 	public String createNewProject(final Model model, @ModelAttribute Project project,
-			@RequestParam(required = false, defaultValue = "false") boolean useCartSamples) {
+			@RequestParam(required = false, defaultValue = "false") boolean useCartSamples,
+			@RequestParam(name = "lockSamples", required = false, defaultValue = "true") boolean owner) {
 
 		try {
 			if (useCartSamples) {
 				Map<Project, List<Sample>> selected = cartController.getSelected();
 
-				List<Long> sampleIds = selected.entrySet().stream().flatMap(e -> e.getValue().stream().filter(s -> {
-					return canModifySample(s);
-				}).map(i -> i.getId())).collect(Collectors.toList());
+				List<Long> sampleIds = selected.entrySet()
+						.stream()
+						.flatMap(e -> e.getValue()
+								.stream()
+								.filter(s -> {
+									return canModifySample(s);
+								})
+								.map(i -> i.getId()))
+						.collect(Collectors.toList());
 
-				project = projectService.createProjectWithSamples(project, sampleIds);
+				project = projectService.createProjectWithSamples(project, sampleIds, owner);
 			} else {
 				project = projectService.create(project);
 			}
 		} catch (ConstraintViolationException e) {
 			model.addAttribute("errors", getErrorsFromViolationException(e));
 			model.addAttribute("project", project);
-			return getCreateProjectPage(useCartSamples, model);
+			return getCreateProjectPage(useCartSamples, model, owner);
 		}
 
 		return "redirect:/projects/" + project.getId() + "/metadata";
@@ -362,13 +358,9 @@ public class ProjectsController {
 	/**
 	 * Returns the name of a page to add users to a *new* project.
 	 *
-	 * @param model
-	 * 		{@link Model}
-	 * @param principal
-	 * 		a reference to the logged in user.
-	 * @param projectId
-	 * 		the id of the project to find the metadata for.
-	 *
+	 * @param model     {@link Model}
+	 * @param principal a reference to the logged in user.
+	 * @param projectId the id of the project to find the metadata for.
 	 * @return The name of the add users to new project page.
 	 */
 	@RequestMapping("/projects/{projectId}/metadata")
@@ -384,12 +376,9 @@ public class ProjectsController {
 	/**
 	 * Get the page for analyses shared with a given {@link Project}
 	 *
-	 * @param projectId
-	 *            the ID of the {@link Project}
-	 * @param principal
-	 *            the logged in user
-	 * @param model
-	 *            model for view variables
+	 * @param projectId the ID of the {@link Project}
+	 * @param principal the logged in user
+	 * @param model     model for view variables
 	 * @return name of the analysis view page
 	 */
 	@RequestMapping("/projects/{projectId}/analyses")
@@ -408,12 +397,9 @@ public class ProjectsController {
 	/**
 	 * Get the page for analysis output files shared with a given {@link Project}
 	 *
-	 * @param projectId
-	 *            the ID of the {@link Project}
-	 * @param principal
-	 *            the logged in user
-	 * @param model
-	 *            model for view variables
+	 * @param projectId the ID of the {@link Project}
+	 * @param principal the logged in user
+	 * @param model     model for view variables
 	 * @return name of the analysis view page
 	 */
 	@RequestMapping("/projects/{projectId}/analyses/shared-outputs")
@@ -430,12 +416,9 @@ public class ProjectsController {
 	/**
 	 * Get the page for automated analysis output files shared with a given {@link Project}
 	 *
-	 * @param projectId
-	 *            the ID of the {@link Project}
-	 * @param principal
-	 *            the logged in user
-	 * @param model
-	 *            model for view variables
+	 * @param projectId the ID of the {@link Project}
+	 * @param principal the logged in user
+	 * @param model     model for view variables
 	 * @return name of the analysis view page
 	 */
 	@RequestMapping("/projects/{projectId}/analyses/automated-outputs")
@@ -458,11 +441,13 @@ public class ProjectsController {
 	 * @return name of the project edit view
 	 */
 	@RequestMapping(value = "/projects/{projectId}/metadata/edit", method = RequestMethod.GET)
-	public String getProjectMetadataEditPage(final Model model, final Principal principal, @PathVariable long projectId) {
+	public String getProjectMetadataEditPage(final Model model, final Principal principal,
+			@PathVariable long projectId) {
 		Project project = projectService.read(projectId);
 		User user = userService.getUserByUsername(principal.getName());
-		if (user.getSystemRole().equals(Role.ROLE_ADMIN)
-				|| projectService.userHasProjectRole(user, project, ProjectRole.PROJECT_OWNER)) {
+		if (user.getSystemRole()
+				.equals(Role.ROLE_ADMIN) || projectService.userHasProjectRole(user, project,
+				ProjectRole.PROJECT_OWNER)) {
 			if (!model.containsAttribute("errors")) {
 				model.addAttribute("errors", new HashMap<>());
 			}
@@ -531,9 +516,7 @@ public class ProjectsController {
 	 * <p>
 	 * Note: If the search term was not included in the results, it will be added as an option
 	 *
-	 * @param searchTerm
-	 * 		The term to find taxa for
-	 *
+	 * @param searchTerm The term to find taxa for
 	 * @return A {@code List<Map<String,Object>>} which will contain a taxonomic tree of matching terms
 	 */
 	@RequestMapping("/projects/ajax/taxonomy/search")
@@ -560,57 +543,50 @@ public class ProjectsController {
 	}
 
 	/**
-	 *  User mapping to get a list of all project they are on.
+	 * User mapping to get a list of all project they are on.
 	 *
-	 * @param params
-	 * 		{@link DataTablesParams} passed from the UI DataTables instance.
-	 *
+	 * @param params {@link DataTablesParams} passed from the UI DataTables instance.
 	 * @return {@link DataTablesResponse}
 	 */
 	@RequestMapping("/projects/ajax/list")
 	@ResponseBody
 	public DataTablesResponse getAjaxProjectList(@DataTablesRequest DataTablesParams params) {
-		final Page<Project> page = projectService
-				.findProjectsForUser(params.getSearchValue(), params.getCurrentPage(), params.getLength(),
-						params.getSort());
-		List<DataTablesResponseModel> projects = page.getContent().stream().map(this::createDataTablesProject).collect(Collectors.toList());
+		final Page<Project> page = projectService.findProjectsForUser(params.getSearchValue(), params.getCurrentPage(),
+				params.getLength(), params.getSort());
+		List<DataTablesResponseModel> projects = page.getContent()
+				.stream()
+				.map(this::createDataTablesProject)
+				.collect(Collectors.toList());
 		return new DataTablesResponse(params, page, projects);
 	}
 
 	/**
 	 * Admin mapping to get a list of all project they are on.
 	 *
-	 * @param params
-	 * 		{@link DataTablesParams} passed from the UI DataTables instance.
-	 *
+	 * @param params {@link DataTablesParams} passed from the UI DataTables instance.
 	 * @return {@link DataTablesResponse}
 	 */
 	@RequestMapping("/projects/admin/ajax/list")
 	@ResponseBody
 	public DataTablesResponse getAjaxAdminProjectsList(@DataTablesRequest DataTablesParams params) {
-		final Page<Project> page = projectService
-				.findAllProjects(params.getSearchValue(), params.getCurrentPage(), params.getLength(),
-						params.getSort());
-		List<DataTablesResponseModel> projects = page.getContent().stream().map(this::createDataTablesProject).collect(Collectors.toList());
+		final Page<Project> page = projectService.findAllProjects(params.getSearchValue(), params.getCurrentPage(),
+				params.getLength(), params.getSort());
+		List<DataTablesResponseModel> projects = page.getContent()
+				.stream()
+				.map(this::createDataTablesProject)
+				.collect(Collectors.toList());
 		return new DataTablesResponse(params, page, projects);
 	}
 
 	/**
 	 * Export Projects table as either an excel file or CSV
 	 *
-	 * @param type
-	 * 		of file to export (csv or excel)
-	 * @param isAdmin
-	 * 		if the currently logged in user is an administrator
-	 * @param response
-	 * 		{@link HttpServletResponse}
-	 * @param principal
-	 * 		{@link Principal}
-	 * @param locale
-	 * 		{@link Locale}
-	 *
-	 * @throws IOException
-	 * 		thrown if cannot open the {@link HttpServletResponse} {@link OutputStream}
+	 * @param type      of file to export (csv or excel)
+	 * @param isAdmin   if the currently logged in user is an administrator
+	 * @param response  {@link HttpServletResponse}
+	 * @param principal {@link Principal}
+	 * @param locale    {@link Locale}
+	 * @throws IOException thrown if cannot open the {@link HttpServletResponse} {@link OutputStream}
 	 */
 	@RequestMapping("/projects/ajax/export")
 	public void exportProjectsToFile(@RequestParam(value = "dtf") String type,
@@ -630,19 +606,23 @@ public class ProjectsController {
 		// If on the users projects page, give the user their projects.
 		else {
 			User user = userService.getUserByUsername(principal.getName());
-			projects = projectService.getProjectsForUser(user).stream().map(Join::getSubject)
+			projects = projectService.getProjectsForUser(user)
+					.stream()
+					.map(Join::getSubject)
 					.collect(Collectors.toList());
 		}
 
-		List<DTProject> dtProjects = projects.stream().map(this::createDataTablesProject).collect(Collectors.toList());
-		List<String> headers = ImmutableList.of("id", "name", "organism", "samples", "created", "modified").stream()
+		List<DTProject> dtProjects = projects.stream()
+				.map(this::createDataTablesProject)
+				.collect(Collectors.toList());
+		List<String> headers = ImmutableList.of("id", "name", "organism", "samples", "created", "modified")
+				.stream()
 				.map(h -> messageSource.getMessage("projects.table." + h, new Object[] {}, locale))
 				.collect(Collectors.toList());
 
 		// Create the filename
 		Date date = new Date();
-		DateFormat fileDateFormat = new SimpleDateFormat(
-				messageSource.getMessage("date.iso-8601", null, locale));
+		DateFormat fileDateFormat = new SimpleDateFormat(messageSource.getMessage("date.iso-8601", null, locale));
 		String filename = "IRIDA_projects_" + fileDateFormat.format(date);
 
 		response.setHeader("Content-Disposition", "attachment; filename=\"" + filename + "." + type + "\"");
@@ -656,22 +636,17 @@ public class ProjectsController {
 	/**
 	 * Write the projects as a CSV file
 	 *
-	 * @param headers
-	 * 		{@link List} for {@link String} headers for the information.
-	 * @param projects
-	 * 		{@link List} of {@link DTProject} to export
-	 * @param locale
-	 * 		{@link Locale}
-	 * @param response
-	 * 		{@link HttpServletResponse}
-	 *
-	 * @throws IOException
-	 * 		Thrown if cannot get the {@link PrintWriter} for the response
+	 * @param headers  {@link List} for {@link String} headers for the information.
+	 * @param projects {@link List} of {@link DTProject} to export
+	 * @param locale   {@link Locale}
+	 * @param response {@link HttpServletResponse}
+	 * @throws IOException Thrown if cannot get the {@link PrintWriter} for the response
 	 */
 	private void writeProjectsToCsvFile(List<String> headers, List<DTProject> projects, Locale locale,
 			HttpServletResponse response) throws IOException {
 		PrintWriter writer = response.getWriter();
-		try(CSVPrinter printer = new CSVPrinter(writer, CSVFormat.DEFAULT.withRecordSeparator(System.lineSeparator()))) {
+		try (CSVPrinter printer = new CSVPrinter(writer,
+				CSVFormat.DEFAULT.withRecordSeparator(System.lineSeparator()))) {
 			printer.printRecord(headers);
 
 			DateFormat dateFormat = new SimpleDateFormat(messageSource.getMessage("locale.date.long", null, locale));
@@ -692,17 +667,11 @@ public class ProjectsController {
 	/**
 	 * Write the projects as a Excel file
 	 *
-	 * @param headers
-	 * 		{@link List} for {@link String} headers for the information.
-	 * @param projects
-	 * 		{@link List} of {@link DTProject} to export
-	 * @param locale
-	 * 		{@link Locale}
-	 * @param response
-	 * 		{@link HttpServletResponse}
-	 *
-	 * @throws IOException
-	 * 		Thrown if cannot get the {@link OutputStream} for the response
+	 * @param headers  {@link List} for {@link String} headers for the information.
+	 * @param projects {@link List} of {@link DTProject} to export
+	 * @param locale   {@link Locale}
+	 * @param response {@link HttpServletResponse}
+	 * @throws IOException Thrown if cannot get the {@link OutputStream} for the response
 	 */
 	private void writeProjectsToExcelFile(List<String> headers, List<DTProject> projects, Locale locale,
 			HttpServletResponse response) throws IOException {
@@ -722,16 +691,22 @@ public class ProjectsController {
 		for (DTProject p : projects) {
 			Row row = sheet.createRow(rowCount++);
 			int cellCount = 0;
-			row.createCell(cellCount++).setCellValue(String.valueOf(p.getId()));
-			row.createCell(cellCount++).setCellValue(p.getName());
-			row.createCell(cellCount++).setCellValue(p.getOrganism());
-			row.createCell(cellCount++).setCellValue(String.valueOf(p.getSamples()));
-			row.createCell(cellCount++).setCellValue(dateFormat.format(p.getCreatedDate()));
-			row.createCell(cellCount).setCellValue(dateFormat.format(p.getModifiedDate()));
+			row.createCell(cellCount++)
+					.setCellValue(String.valueOf(p.getId()));
+			row.createCell(cellCount++)
+					.setCellValue(p.getName());
+			row.createCell(cellCount++)
+					.setCellValue(p.getOrganism());
+			row.createCell(cellCount++)
+					.setCellValue(String.valueOf(p.getSamples()));
+			row.createCell(cellCount++)
+					.setCellValue(dateFormat.format(p.getCreatedDate()));
+			row.createCell(cellCount)
+					.setCellValue(dateFormat.format(p.getModifiedDate()));
 		}
 
 		// Write the file
-		try(OutputStream stream = response.getOutputStream()) {
+		try (OutputStream stream = response.getOutputStream()) {
 			workbook.write(stream);
 			stream.flush();
 		}
@@ -740,16 +715,15 @@ public class ProjectsController {
 	/**
 	 * Changes a {@link ConstraintViolationException} to a usable map of strings for displaing in the UI.
 	 *
-	 * @param e
-	 * 		{@link ConstraintViolationException} for the form submitted.
-	 *
+	 * @param e {@link ConstraintViolationException} for the form submitted.
 	 * @return Map of string {fieldName, error}
 	 */
 	private Map<String, String> getErrorsFromViolationException(ConstraintViolationException e) {
 		Map<String, String> errors = new HashMap<>();
 		for (ConstraintViolation<?> violation : e.getConstraintViolations()) {
 			String message = violation.getMessage();
-			String field = violation.getPropertyPath().toString();
+			String field = violation.getPropertyPath()
+					.toString();
 			errors.put(field, message);
 		}
 		return errors;
@@ -770,12 +744,12 @@ public class ProjectsController {
 	/**
 	 * Test whether the logged in user can modify a {@link Sample}
 	 *
-	 * @param sample
-	 *            the {@link Sample} to check
+	 * @param sample the {@link Sample} to check
 	 * @return true if they can modify
 	 */
 	private boolean canModifySample(Sample sample) {
-		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		Authentication authentication = SecurityContextHolder.getContext()
+				.getAuthentication();
 
 		return updateSamplePermission.isAllowed(authentication, sample);
 	}
@@ -785,16 +759,15 @@ public class ProjectsController {
 	 * <p>
 	 * /** Recursively transform a {@link TreeNode} into a json parsable map object
 	 *
-	 * @param node
-	 * 		The node to transform
-	 *
+	 * @param node The node to transform
 	 * @return A Map<String,Object> which may contain more children
 	 */
 	private Map<String, Object> transformTreeNode(TreeNode<String> node) {
 		Map<String, Object> current = new HashMap<>();
 
 		// add the node properties to the map
-		for (Entry<String, Object> property : node.getProperties().entrySet()) {
+		for (Entry<String, Object> property : node.getProperties()
+				.entrySet()) {
 			current.put(property.getKey(), property.getValue());
 		}
 
@@ -817,9 +790,7 @@ public class ProjectsController {
 	/**
 	 * Extract the details of the a {@link Project} into a {@link DTProject} which is consumable by the UI
 	 *
-	 * @param project
-	 * 		{@link Project}
-	 *
+	 * @param project {@link Project}
 	 * @return {@link DTProject}
 	 */
 	private DTProject createDataTablesProject(Project project) {